[Core] Move taxes from product to variant
