+++ src/Sylius/Bundle/CoreBundle/Resources/config/doctrine/model/Product.orm.xml
@@ -21,11 +21,6 @@
 
         <field name="variantSelectionMethod" column="variant_selection_method" type="string" nullable="false" />
 
-        <many-to-one field="taxCategory" target-entity="Sylius\Component\Taxation\Model\TaxCategoryInterface">
-            <join-column name="tax_category_id" referenced-column-name="id" nullable="true" on-delete="SET NULL" />
-            <gedmo:versioned />
-        </many-to-one>
-
         <many-to-one field="shippingCategory" target-entity="Sylius\Component\Shipping\Model\ShippingCategoryInterface">
             <join-column name="shipping_category_id" referenced-column-name="id" nullable="true" on-delete="SET NULL" />
             <gedmo:versioned />
