+++ src/Sylius/Bundle/CoreBundle/Resources/config/doctrine/model/ProductVariant.orm.xml
@@ -63,6 +63,11 @@
             </cascade>
         </one-to-many>
 
+        <many-to-one field="taxCategory" target-entity="Sylius\Component\Taxation\Model\TaxCategoryInterface">
+            <join-column name="tax_category_id" referenced-column-name="id" nullable="true" on-delete="SET NULL" />
+            <gedmo:versioned />
+        </many-to-one>
+
         <gedmo:loggable />
     </mapped-superclass>
