+++ src/Sylius/Bundle/WebBundle/Resources/views/Backend/Product/show.html.twig
@@ -134,12 +134,12 @@
                     </td>
                 </tr>
                 <tr>
-                    <td>{{ 'sylius.product.tax_category'|trans }}</td>
-                    <td>{{ product.taxCategory|default('sylius.product.no_tax_category'|trans) }}</td>
-                </tr>
-                <tr>
                     <td>{{ 'sylius.product.shipping_category'|trans }}</td>
                     <td>{{ product.shippingCategory|default('sylius.product.no_shipping_category'|trans) }}</td>
+                </tr>
+                <tr>
+                    <td>{{ 'sylius.product.tax_category'|trans }}</td>
+                    <td>{{ product.masterVariant.taxCategory|default('sylius.product.no_tax_category'|trans) }}</td>
                 </tr>
                 <tr>
                     <td>{{ 'sylius.product.available_on_demand'|trans }}</td>
