+++ src/Sylius/Bundle/WebBundle/Resources/views/Backend/ProductVariant/_form.html.twig
@@ -33,6 +33,7 @@
         <hr />
         {{ form_row(form.availableOnDemand) }}
         {{ form_row(form.onHand) }}
+        {{ form_row(form.taxCategory) }}
     </div>
 
     <div class="col-md-4">
