+++ src/Sylius/Component/Core/Model/Product.php
@@ -17,7 +17,6 @@ use Sylius\Component\Addressing\Model\ZoneInterface;
 use Sylius\Component\Channel\Model\ChannelInterface as BaseChannelInterface;
 use Sylius\Component\Product\Model\Product as BaseProduct;
 use Sylius\Component\Shipping\Model\ShippingCategoryInterface;
-use Sylius\Component\Taxation\Model\TaxCategoryInterface;
 use Sylius\Component\Taxonomy\Model\TaxonInterface as BaseTaxonInterface;
 
 /**
@@ -28,43 +27,26 @@ use Sylius\Component\Taxonomy\Model\TaxonInterface as BaseTaxonInterface;
 class Product extends BaseProduct implements ProductInterface
 {
     /**
-     * Variant selection method.
-     *
      * @var string
      */
     protected $variantSelectionMethod;
 
     /**
-     * Taxons.
-     *
      * @var Collection|BaseTaxonInterface[]
      */
     protected $taxons;
 
     /**
-     * Tax category.
-     *
-     * @var TaxCategoryInterface
-     */
-    protected $taxCategory;
-
-    /**
-     * Shipping category.
-     *
      * @var ShippingCategoryInterface
      */
     protected $shippingCategory;
 
     /**
-     * Not allowed to ship in this zone.
-     *
      * @var ZoneInterface
      */
     protected $restrictedZone;
 
     /**
-     * Channels in which this product is available.
-     *
      * @var ChannelInterface[]|Collection
      */
     protected $channels;
@@ -74,9 +56,6 @@ class Product extends BaseProduct implements ProductInterface
      */
     protected $mainTaxon;
 
-    /**
-     * Constructor.
-     */
     public function __construct()
     {
         parent::__construct();
@@ -238,24 +217,6 @@ class Product extends BaseProduct implements ProductInterface
     /**
      * {@inheritdoc}
      */
-    public function getTaxCategory()
-    {
-        return $this->taxCategory;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function setTaxCategory(TaxCategoryInterface $category = null)
-    {
-        $this->taxCategory = $category;
-
-        return $this;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
     public function getShippingCategory()
     {
         return $this->shippingCategory;
@@ -376,8 +337,6 @@ class Product extends BaseProduct implements ProductInterface
     public function setShortDescription($shortDescription)
     {
         $this->translate()->setShortDescription($shortDescription);
-
-        return $this;
     }
 
     /**
